Added Flocking/grouping and separation
Basically every cell now has a dna property, used to share food and poison to the neighbors in his range.
Every time they eat, the radius increases to help feeding more cells.

It's sketchy, it's definitely not perfect, but it kinda works.
Needs definitely an improvement but it's a starting point.